+++ students/21K0647/21K0647-p02/src/main/java/ru/mirea/practice3/s21k0647/Circle.java
@@ -1,18 +1,18 @@
-package ru.mirea.practice3.s21K0647;
+package ru.mirea.practice3.s21k0647;
 
-public class Point {
-    double x;
-    double y;
+public class Circle extends Point {
 
-    public Point() {
-        x = 0;
-        y = 0;
+    Point tT;
+    double r;
+
+    public Circle() {
+        r = 0;
     }
 
-    public Point(double x, double y) {
-        this.x = x;
-        this.y = y;
+    public Circle(double r) {
+        this.r = r;
     }
+
 
     public double getX() {
         return x;
@@ -29,4 +29,5 @@ public class Point {
     public void setY(double y) {
         this.y = y;
     }
+
 }
