+++ students/21K0647/21K0647-p02/src/main/java/ru/mirea/practice7/s21k0647/Book.java
@@ -1,4 +1,4 @@
-package ru.mirea.practice7.s21K0647;
+package ru.mirea.practice7.s21k0647;
 
 public class Book {
     String author;
