+++ students/21K0647/21K0647-p02/src/main/java/ru/mirea/practice7/s21k0647/Bookshelf.java
@@ -1,4 +1,4 @@
-package ru.mirea.practice7.s21K0647;
+package ru.mirea.practice7.s21k0647;
 
 import java.util.ArrayList;
