+++ students/21K0647/21K0647-p02/src/main/java/ru/mirea/practice8/s21k0647/Permutation.java
@@ -1,4 +1,4 @@
-package ru.mirea.practice8.s21K0647;
+package ru.mirea.practice8.s21k0647;
 
 import java.util.Scanner;
